+++ package.json
@@ -2,7 +2,6 @@
   "name": "sd-0x-block15-project-frontend-online-store",
   "version": "0.1.0",
   "private": true,
-  "proxy": "https://api.mercadolibre.com/sites/MLB/categories",
   "dependencies": {
     "@testing-library/jest-dom": "^5.5.0",
     "@testing-library/react": "^10.0.4",
