+++ src/components/Aside/style.css
@@ -1,0 +1,4 @@
+.categories {
+  display: flex;
+  flex-direction: column;
+}
