+++ src/services/api.js
@@ -1,12 +1,10 @@
 const api = 'https://api.mercadolibre.com/sites/MLB/categories';
 
 export const getCategories = async () => {
-  const response = await fetch(api, { mode:'cors' });
+  const response = await fetch(api, { mode: 'cors' });
   const data = await response.json();
   return data;
 };
-
-
 
 export async function getProductsFromCategoryAndQuery(categoryId, query) {
   const data = await fetch(
